extract shape generator functionality to own module
